requirements.txt
streamlit
pyyaml
pytest
